requirements.txt
Django==1.8.5
wsgiref==0.1.2

django-background-tasks==1.0.7
django-compat==1.0.7

honcho==0.5.0

pusher==1.2.3
cffi==1.2.1
cryptography==1.0.2
enum34==1.0.4
idna==2.0
ipaddress==1.0.14
ndg-httpsclient==0.4.0
pyOpenSSL==0.15.1
pyasn1==0.1.9
pycparser==2.14
requests==2.8.0
six==1.10.0
urllib3==1.12
